Muita coisa
Portal Umadjaf
Área de Login e Cadastro:
Cadastro
Para realizar o cadastro, o usuário deve informar os seguintes dados:
Nome
Data de Nascimento
Número de Telefone
Congregação
Indicação se pertence ou não à Umadjaf
Senha
Login
Para realizar o login, o usuário deve informar:
Número de Telefone
Senha
Perfil:
Inicialmente, o perfil é oculto para usuários não confirmados como membros da Umadjaf. Um perfil alternativo será criado para não membros.
Após a confirmação, os membros terão acesso às seguintes funcionalidades:
Criação de devocionais
Visualização dos devocionais criados em seu perfil
Visualização de fotos de eventos em que foram marcados ou solicitaram marcação
Obs: A área de pedidos será incluída junto com a nova atualização para a implementação da loja.
Devocionais:
Tela Inicial
Exibição dos 3 últimos devocionais verificados.
Criação de Devocionais
Todos os usuários podem criar devocionais.
Devocionais verificados ou com o selo azul são revisados e podem aparecer na frente.
Devocionais não verificados podem ser apagados.
Área de Verificação
Exibição de devocionais verificados e não verificados, distinguidos pelo selo.
Pesquisa de devocionais pode ser feita pelo texto, título e referência bíblica.
Requisitos para Criação de Devocionais
Título
Referência Bíblica
Texto
Banner (idealmente na proporção 32:9)
Eventos:
Eventos podem ser registrados por coordenadores e administradores.
Administradores podem marcar se o evento é geral ou local.
Coordenadores podem registrar apenas eventos locais.
Eventos gerais aparecem na tela inicial.
Eventos locais aparecem somente na aba de eventos.
Requisitos para Banners de Eventos
Composição por dois itens: fundo e logo, para uma melhor fluidez do site.

+++ manager/views.py
@@ -1,19 +1,21 @@
-from django.shortcuts import render, get_object_or_404, redirect
 from django.http import HttpResponseRedirect
+from django.shortcuts import get_object_or_404, redirect, render
 from django.urls import reverse
-from profiles.models import User, UserRoles, Roles, IsUmadjaf
+from users.models import IsUmadjaf, Roles, User, UserRoles
+from .forms import (CalendarForm, CongregationForm, IsUmadjafForm, UserForm,
+                    UserRolesForm)
 from .models import Congregations
-from .forms import UserRolesForm, IsUmadjafForm, UserForm, CongregationForm
+from events.models import Event
+from articles.models import Articles
 
 
 # Funções do novo painel de controle
-
 def adm_panel(request):
     is_authenticated = request.user.is_authenticated  # Verifica se o usuário está autenticado
-    user_is_admin = request.user.is_staff and request.user.is_superuser # Verifica se o usuário é admin
+    user_is_admin = request.user.is_staff # Verifica se o usuário é admin
     if not is_authenticated:  # Se não estiver autenticado, redireciona para a página de login
         if not user_is_admin:
-            return redirect('profiles:login')
+            return redirect('users:login')
     return render(
         request,
         'manager/pages/adm_panel.html',
@@ -26,10 +28,10 @@ def adm_panel(request):
 # Funções de gerenciamento de usuários
 def users(request):
     is_authenticated = request.user.is_authenticated  # Verifica se o usuário está autenticado
-    user_is_admin = request.user.is_staff and request.user.is_superuser # Verifica se o usuário é admin
+    user_is_admin = request.user.is_staff # Verifica se o usuário é admin
     if not is_authenticated:  # Se não estiver autenticado, redireciona para a página de login
         if user_is_admin:
-            return redirect('profiles:login')
+            return redirect('users:login')
     
     users = User.objects.all()  # Busca todos os usuários cadastrados
     roles = Roles.objects.all()  # Busca todos os cargos cadastrados
@@ -45,6 +47,9 @@ def users(request):
 
 
 def delete_user(request, user_id):
+    if user_id == 1:
+        return HttpResponseRedirect(reverse('manager:users'))
+    
     user = get_object_or_404(User, pk=user_id)  # Busca o usuário pelo ID
     user.delete()  # Deleta o usuário
     return HttpResponseRedirect(reverse('manager:users'))
@@ -52,10 +57,10 @@ def delete_user(request, user_id):
 
 def user_edit(request, user_id):
     is_authenticated = request.user.is_authenticated  # Verifica se o usuário está autenticado
-    user_is_admin = request.user.is_staff and request.user.is_superuser # Verifica se o usuário é admin
+    user_is_admin = request.user.is_staff # Verifica se o usuário é admin
     if not is_authenticated:  # Se não estiver autenticado, redireciona para a página de login
         if not user_is_admin:
-            return redirect('profiles:login')
+            return redirect('users:login')
     user = User.objects.get(id=user_id)  # Busca o usuário pelo ID
     user_roles = UserRoles.objects.get(user_id=user_id)  # Busca o cargo do usuário pelo ID
     umadjaf = IsUmadjaf.objects.get(user_id=user_id)  # Busca se o usuário é umadjaf pelo ID
@@ -68,6 +73,8 @@ def user_edit(request, user_id):
         print("Log: Formulários recebidos")
         if user_form.is_valid() and user_roles_form.is_valid() and umadjaf_form.is_valid():
             print("Log: Formulários válidos")
+            if user_id == 1:
+                return HttpResponseRedirect(reverse('manager:users'))
             user_form.save()
             user_roles_form.save()
             umadjaf_form.save()
@@ -92,10 +99,10 @@ def user_edit(request, user_id):
 # Funções de gerenciamento de cargos
 def users_roles(request):
     is_authenticated = request.user.is_authenticated  # Verifica se o usuário está autenticado
-    user_is_admin = request.user.is_staff and request.user.is_superuser # Verifica se o usuário é admin
+    user_is_admin = request.user.is_staff # Verifica se o usuário é admin
     if not is_authenticated:  # Se não estiver autenticado, redireciona para a página de login
         if not user_is_admin:
-            return redirect('profiles:login')
+            return redirect('users:login')
 
     users_roles_all = UserRoles.objects.all()
     return render(
@@ -110,15 +117,17 @@ def users_roles(request):
 
 def role_changer(request, user_id):
     is_authenticated = request.user.is_authenticated  # Verifica se o usuário está autenticado
-    user_is_admin = request.user.is_staff and request.user.is_superuser # Verifica se o usuário é admin
+    user_is_admin = request.user.is_staff # Verifica se o usuário é admin
     if not is_authenticated:  # Se não estiver autenticado, redireciona para a página de login
         if not user_is_admin:
-            return redirect('profiles:login')
+            return redirect('users:login')
 
     user_role = UserRoles.objects.get(user_id=user_id)
     if request.method == 'POST':
         user_role_form = UserRolesForm(request.POST, instance=user_role)
         if user_role_form.is_valid():
+            if user_id == 1:
+                return HttpResponseRedirect(reverse('manager:users_roles'))
             user_role_form.save()
             return redirect('manager:users_roles')
     else:
@@ -135,10 +144,10 @@ def role_changer(request, user_id):
 # Funções de gerenciamento de congregações
 def congregations(request):
     is_authenticated = request.user.is_authenticated  # Verifica se o usuário está autenticado
-    user_is_admin = request.user.is_staff and request.user.is_superuser # Verifica se o usuário é admin
+    user_is_admin = request.user.is_staff # Verifica se o usuário é admin
     if not is_authenticated:  # Se não estiver autenticado, redireciona para a página de login
         if not user_is_admin:
-            return redirect('profiles:login')
+            return redirect('users:login')
 
     congregations = Congregations.objects.all()
     return render(
@@ -153,10 +162,10 @@ def congregations(request):
 
 def congregation_add(request):
     is_authenticated = request.user.is_authenticated  # Verifica se o usuário está autenticado
-    user_is_admin = request.user.is_staff and request.user.is_superuser # Verifica se o usuário é admin
+    user_is_admin = request.user.is_staff # Verifica se o usuário é admin
     if not is_authenticated:  # Se não estiver autenticado, redireciona para a página de login
         if not user_is_admin:
-            return redirect('profiles:login')
+            return redirect('users:login')
 
     if request.method == 'POST':
         congregation = Congregations(
@@ -196,10 +205,10 @@ def congregation_delete(request, congregation_id):
 # Funções de gerenciamento de Membros
 def members(request):
     is_authenticated = request.user.is_authenticated  # Verifica se o usuário está autenticado
-    user_is_admin = request.user.is_staff and request.user.is_superuser # Verifica se o usuário é admin
+    user_is_admin = request.user.is_staff # Verifica se o usuário é admin
     if not is_authenticated:  # Se não estiver autenticado, redireciona para a página de login
         if not user_is_admin:
-            return redirect('profiles:login')
+            return redirect('users:login')
 
     members = IsUmadjaf.objects.all()
     return render(
@@ -219,6 +228,9 @@ def member_positive(request, member_id):
     member.is_umadjaf = False
     umadjaf.checked = True
 
+    if member_id == 1:
+        return HttpResponseRedirect(reverse('manager:members'))
+
     member.save()
     umadjaf.save()
 
@@ -226,14 +238,109 @@ def member_positive(request, member_id):
 
 
 def member_negative(request, member_id):
-    print(member_id)
     member = User.objects.get(id=member_id)
     umadjaf = IsUmadjaf.objects.get(user_id=member)
 
     member.is_umadjaf = False
     umadjaf.checked = False
 
+    if member_id == 1:
+        return HttpResponseRedirect(reverse('manager:members'))
+
     member.save()
     umadjaf.save()
 
     return HttpResponseRedirect(reverse('manager:members'))
+
+
+# Sistema de calendário
+def calendar(request):
+    is_authenticated = request.user.is_authenticated  # Verifica se o usuário está autenticado
+    user_is_admin = request.user.is_staff # Verifica se o usuário é admin
+    if not is_authenticated:  # Se não estiver autenticado, redireciona para a página de login
+        if not user_is_admin:
+            return redirect('users:login')
+
+    calendar = Event.objects.all()
+
+    return render(
+        request,
+        'manager/pages/calendar/calendar.html',
+        context={
+            'is_authenticated': is_authenticated,
+            'calendar_all': calendar
+        }
+    )
+
+
+def calendar_edit(request, calendar_id):
+    calendar = get_object_or_404(Event, pk=calendar_id)
+    event_logo = calendar.logo
+    event_background = calendar.background
+    if request.method == 'POST':
+        calendar_edit_form = CalendarForm(request.POST, request.FILES, instance=calendar)
+        if calendar_edit_form.is_valid():
+            # O objeto 'calendar' será atualizado com os dados do formulário
+            calendar_edit_form.save()
+            return redirect('manager:calendar')
+    else:
+        calendar_edit_form = CalendarForm(instance=calendar)
+
+    return render(
+        request,
+        'manager/pages/calendar/partials/calendar_edit.html',
+        context={
+            'calendar_edit_form': calendar_edit_form,
+            'calendar_id': calendar_id,
+            'event_logo': event_logo,
+            'event_background': event_background
+        }
+    )
+
+
+def calendar_delete(request, calendar_id):
+    calendar = get_object_or_404(Event, pk=calendar_id)
+    calendar.delete()
+    return HttpResponseRedirect(reverse('manager:calendar'))
+
+
+# Funções de gerenciamento de Artigos
+def articles(request):
+    is_authenticated = request.user.is_authenticated  # Verifica se o usuário está autenticado
+    user_is_admin = request.user.is_staff # Verifica se o usuário é admin
+    articles = Articles.objects.all()
+
+    if not is_authenticated:  # Se não estiver autenticado, redireciona para a página de login
+        if not user_is_admin:
+            return redirect('users:login')
+
+    return render(
+        request,
+        'manager/pages/articles/articles.html',
+        context={
+            'is_authenticated': is_authenticated,
+            'articles_all': articles
+        }
+    )
+
+
+def article_verify(request, article_id):
+    article = Articles.objects.get(id=article_id)
+    article.is_official = True
+    article.post_unlock = True
+    article.save()
+    return HttpResponseRedirect(reverse('manager:articles'))
+
+
+def article_unverify(request, article_id):
+    article = Articles.objects.get(id=article_id)
+    article.is_official = False
+    article.post_unlock = False
+    article.save()
+    return HttpResponseRedirect(reverse('manager:articles'))
+
+
+def article_delete(request, article_id):
+    article = get_object_or_404(Articles, pk=article_id)
+    article.delete()
+    return HttpResponseRedirect(reverse('manager:articles'))
